Replace Context with Zustand
