+++ package.json
@@ -36,11 +36,12 @@
     "next": "15.3.0",
     "nprogress": "^0.2.0",
     "postcss": "^8.5.3",
-    "react": "18.3.1",
-    "react-dom": "18.3.1",
+    "react": "19.1.0",
+    "react-dom": "19.1.0",
     "react-hook-form": "^7.55.0",
     "react-instantsearch-dom": "^6.40.4",
-    "uuid": "^11.1.0"
+    "uuid": "^11.1.0",
+    "zustand": "^5.0.3"
   },
   "devDependencies": {
     "@lhci/cli": "^0.14.0",
@@ -52,7 +53,7 @@
     "@types/uuid": "^10.0.0",
     "@typescript-eslint/eslint-plugin": "^8.29.1",
     "@typescript-eslint/parser": "^8.29.1",
-    "babel-plugin-styled-components": "^2.1.4",   
+    "babel-plugin-styled-components": "^2.1.4",
     "eslint-config-next": "^15.3.0",
     "postcss-preset-env": "^10.1.5",
     "prettier": "^3.5.3",
