+++ src/pages/_app.tsx
@@ -3,8 +3,6 @@ import Router from 'next/router';
 import NProgress from 'nprogress';
 import { ApolloProvider } from '@apollo/client';
 
-// State import
-import { CartProvider } from '@/stores/CartProvider';
 import client from '@/utils/apollo/ApolloClient';
 
 // Types
@@ -22,9 +20,7 @@ Router.events.on('routeChangeError', () => NProgress.done());
 function MyApp({ Component, pageProps }: AppProps) {
   return (
     <ApolloProvider client={client}>
-      <CartProvider>
-        <Component {...pageProps} />
-      </CartProvider>
+      <Component {...pageProps} />
     </ApolloProvider>
   );
 }
